Move node based settlement submission into own file
